adding singularityce and shared singularity base
